Changes on Class Files
Several changes on class files.
I know that is not the way to commit stuff,
but the principal changes stay in  the PrinterSettingsClass,
that was renamed from PrinterConfigs.
Add a class PrinterSettings to manage the data on
PrinterSettingsWidget.
Add a signal on this class to connect on ManualControlClass,
to manage some widgets.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>