New Class Inherit from QVTKWidget2
Well, since VTK have 2 class QVTKWidget, where one depreciated and
other not, I though that would be better that I work with two
classes, one with the old class and one with the new one, since
that the help that the examples that i found, work if the
old class. So i work with the 2 classes until I pass the
code to the new class.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>