Port the vtkWidget class to use QVTKWidget
Port the inherit from this class.
This class has a deprecated library!
QVTKWidget is disused, so this class will not be used for long.
Only I'm working with her because I'm learning VTK and many Qt + VTK tutorials still use this library,
with time I will be carrying to the new class.
Thank you for your understanding.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>